Create rise-up-and-salute-the-sun.md
Everybody has a little bit of the sun and moon in them. Everybody has a little bit of man, woman, and animal in them. Darks and lights in them. Everyone is part of a connected cosmic system. Part earth and sea, wind and fire, with some salt and dust swimming in them. We have a universe within ourselves that mimics the universe outside. None of us are just black or white, or never wrong and always right. No one. No one exists without polarities. Everybody has good and bad forces working with them, against them, and within them.